tasktable fix, detailed project page
