+++ src/pages/Process/components/TasksTable/TasksTable.css
@@ -68,3 +68,19 @@ td {
 
     height: 5.3vh;
 }
+.no-data{
+    width: 100%;
+    height: 30vh;
+
+    border-radius: 15px;
+    border: 1px dashed #d9d9d9;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 16px;
+
+    color: #818181;
+
+    padding: 4%;
+}
