+++ src/pages/Project/ProjectDetail.css
@@ -1,14 +1,47 @@
+.project-detail-wrapper{
+    display: flex;
+    justify-content: center;
+}
 .project-detail-container{
     width: 90%;
+    max-width: 1300px;
 
     margin: 5% 5% 0 7%;
 
     display: grid;
-    grid-template-columns: repeat(2, 1fr);
+    grid-template-columns: 38% 50%;
+    grid-gap: 2%;
 }
-.project-detail-container:nth-child(1){
+.project-detail-title:nth-child(1){
     grid-column: 1 / span 2;
 }
 .project-detail-title{
+    font-size: 45px;
+}
+.project-detail-time:nth-child(2){
+    grid-column: 1 / span 2;
+    padding-bottom: 3%;
+    margin-bottom: 3%;
+    border-bottom: 1px solid #d6d6d6;
+}
+.project-detail-members{
+    /*border: 1px solid #d9d9d9;*/
+    padding: 3% 0;
+    font-size: 21px;
+}
+.project-detail-description-wrapper{
+    /*border: 1px solid #d9d9d9;*/
+}
+.project-detail-description{
+    font-size: 20px;
+}
+.project-detail-tasks-info-wrapper:nth-child(5){
+    margin-top: 5%;
+    grid-column: 1 / span 2;
+}
+.project-detail-tasks-info-title{
     font-size: 30px;
+}
+.project-detail-tasks-info-poss{
+    font-size: 20px;
 }
